update docs and delete old files
