+++ examples/darcy/darcy2d.py
@@ -1,139 +1,201 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
-#
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
+
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-#
+
 #     http://www.apache.org/licenses/LICENSE-2.0
-#
+
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import paddlescience as psci
 import numpy as np
-import paddle
 
-cfg = psci.utils.parse_args()
+import ppsci
+from ppsci.utils import config
+from ppsci.utils import logger
 
-if cfg is not None:
-    # Geometry
-    npoints = cfg['Geometry']['npoints']
-    seed_num = cfg['Geometry']['seed']
-    sampler_method = cfg['Geometry']['sampler_method']
-    # Network
-    epochs = cfg['Global']['epochs']
-    num_layers = cfg['Model']['num_layers']
-    hidden_size = cfg['Model']['hidden_size']
-    activation = cfg['Model']['activation']
-    # Optimizer
-    learning_rate = cfg['Optimizer']['lr']['learning_rate']
-    # Post-processing
-    solution_filename = cfg['Post-processing']['solution_filename']
-    vtk_filename = cfg['Post-processing']['vtk_filename']
-    checkpoint_path = cfg['Post-processing']['checkpoint_path']
-else:
-    # Config
-    paddle.enable_static()
-    # Geometry
-    npoints = 10201
-    seed_num = 1
-    sampler_method = 'uniform'
-    # Network
-    epochs = 10000
-    num_layers = 5
-    hidden_size = 20
-    activation = 'tanh'
-    # Optimizer
-    learning_rate = 0.001
-    # Post-processing
-    solution_filename = 'output_darcy2d'
-    vtk_filename = 'output_darcy2d'
-    checkpoint_path = 'checkpoints'
+if __name__ == "__main__":
+    args = config.parse_args()
+    # set random seed for reproducibility
+    ppsci.utils.misc.set_random_seed(42)
+    # set output directory
+    OUTPUT_DIR = "./output_darcy2d" if not args.output_dir else args.output_dir
+    # initialize logger
+    logger.init_logger("ppsci", f"{OUTPUT_DIR}/train.log", "info")
 
-paddle.seed(seed_num)
-np.random.seed(seed_num)
+    # set model
+    model = ppsci.arch.MLP(("x", "y"), ("p",), 5, 20, "tanh", False, False)
 
-psci.config.set_dtype("float32")
+    # set equation
+    equation = {"Poisson": ppsci.equation.Poisson(2)}
 
-# ref solution 
-ref_sol = lambda x, y: np.sin(2.0 * np.pi * x) * np.cos(2.0 * np.pi * y)
+    # set geometry
+    geom = {"rect": ppsci.geometry.Rectangle((0.0, 0.0), (1.0, 1.0))}
 
-# ref rhs
-ref_rhs = lambda x, y: 8.0 * np.pi**2 * np.sin(2.0 * np.pi * x) * np.cos(2.0 * np.pi * y)
+    # set dataloader config
+    ITERS_PER_EPOCH = 1
+    train_dataloader_cfg = {
+        "dataset": "IterableNamedArrayDataset",
+        "iters_per_epoch": ITERS_PER_EPOCH,
+    }
 
-# set geometry and boundary
-geo = psci.geometry.Rectangular(origin=(0.0, 0.0), extent=(1.0, 1.0))
+    NPOINT_PDE = 99**2
+    NPOINT_TOP = 101
+    NPOINT_BOTTOM = 101
+    NPOINT_LEFT = 99
+    NPOINT_RIGHT = 99
 
-geo.add_boundary(name="top", criteria=lambda x, y: y == 1.0)
-geo.add_boundary(name="down", criteria=lambda x, y: y == 0.0)
-geo.add_boundary(name="left", criteria=lambda x, y: x == 0.0)
-geo.add_boundary(name="right", criteria=lambda x, y: x == 1.0)
+    # set constraint
+    def poisson_ref_compute_func(_in):
+        return (
+            -8.0
+            * (np.pi**2)
+            * np.sin(2.0 * np.pi * _in["x"])
+            * np.cos(2.0 * np.pi * _in["y"])
+        )
 
-# discretize geometry
-geo_disc = geo.discretize(npoints=npoints, method=sampler_method)
+    pde_constraint = ppsci.constraint.InteriorConstraint(
+        equation["Poisson"].equations,
+        {"poisson": poisson_ref_compute_func},
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_PDE},
+        ppsci.loss.MSELoss("sum"),
+        evenly=True,
+        name="EQ",
+    )
+    bc_top = ppsci.constraint.BoundaryConstraint(
+        {"p": lambda out: out["p"]},
+        {
+            "p": lambda _in: np.sin(2.0 * np.pi * _in["x"])
+            * np.cos(2.0 * np.pi * _in["y"])
+        },
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_TOP},
+        ppsci.loss.MSELoss("sum"),
+        criteria=lambda x, y: np.isclose(y, 1.0),
+        name="BC_top",
+    )
+    bc_bottom = ppsci.constraint.BoundaryConstraint(
+        {"p": lambda out: out["p"]},
+        {
+            "p": lambda _in: np.sin(2.0 * np.pi * _in["x"])
+            * np.cos(2.0 * np.pi * _in["y"])
+        },
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_BOTTOM},
+        ppsci.loss.MSELoss("sum"),
+        criteria=lambda x, y: np.isclose(y, 0.0),
+        name="BC_bottom",
+    )
+    bc_left = ppsci.constraint.BoundaryConstraint(
+        {"p": lambda out: out["p"]},
+        {
+            "p": lambda _in: np.sin(2.0 * np.pi * _in["x"])
+            * np.cos(2.0 * np.pi * _in["y"])
+        },
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_LEFT},
+        ppsci.loss.MSELoss("sum"),
+        criteria=lambda x, y: np.isclose(x, 0.0),
+        name="BC_left",
+    )
+    bc_right = ppsci.constraint.BoundaryConstraint(
+        {"p": lambda out: out["p"]},
+        {
+            "p": lambda _in: np.sin(2.0 * np.pi * _in["x"])
+            * np.cos(2.0 * np.pi * _in["y"])
+        },
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_RIGHT},
+        ppsci.loss.MSELoss("sum"),
+        criteria=lambda x, y: np.isclose(x, 1.0),
+        name="BC_right",
+    )
+    # wrap constraints together
+    constraint = {
+        pde_constraint.name: pde_constraint,
+        bc_top.name: bc_top,
+        bc_bottom.name: bc_bottom,
+        bc_left.name: bc_left,
+        bc_right.name: bc_right,
+    }
 
-# Poisson
-pde = psci.pde.Poisson(dim=2, rhs=ref_rhs)
+    # set training hyper-parameters
+    EPOCHS = 10000 if not args.epochs else args.epochs
 
-# set bounday condition
-bc_top = psci.bc.Dirichlet('u', rhs=ref_sol)
-bc_down = psci.bc.Dirichlet('u', rhs=ref_sol)
-bc_left = psci.bc.Dirichlet('u', rhs=ref_sol)
-bc_right = psci.bc.Dirichlet('u', rhs=ref_sol)
+    # set optimizer
+    optimizer = ppsci.optimizer.Adam(1e-3)((model,))
 
-# add bounday and boundary condition
-pde.add_bc("top", bc_top)
-pde.add_bc("down", bc_down)
-pde.add_bc("left", bc_left)
-pde.add_bc("right", bc_right)
+    # set validator
+    NPOINTS_EVAL = NPOINT_PDE
+    residual_validator = ppsci.validate.GeometryValidator(
+        equation["Poisson"].equations,
+        {"poisson": poisson_ref_compute_func},
+        geom["rect"],
+        {
+            "dataset": "NamedArrayDataset",
+            "total_size": NPOINTS_EVAL,
+            "batch_size": 8192,
+            "sampler": {"name": "BatchSampler"},
+        },
+        ppsci.loss.MSELoss("sum"),
+        evenly=True,
+        metric={"MSE": ppsci.metric.MSE()},
+        name="Residual",
+    )
+    validator = {residual_validator.name: residual_validator}
 
-# discretization pde
-pde_disc = pde.discretize(geo_disc=geo_disc)
+    # set visualizer(optional)
+    # manually collate input data for visualization,
+    NPOINT_BC = NPOINT_TOP + NPOINT_BOTTOM + NPOINT_LEFT + NPOINT_RIGHT
+    vis_points = geom["rect"].sample_interior(NPOINT_PDE + NPOINT_BC, evenly=True)
+    visualizer = {
+        "visulzie_u_v": ppsci.visualize.VisualizerVtu(
+            vis_points,
+            {"p": lambda d: d["p"]},
+            prefix="result_u_v",
+        )
+    }
 
-# Network
-# TODO: remove num_ins and num_outs
-net = psci.network.FCNet(
-    num_ins=2,
-    num_outs=1,
-    num_layers=num_layers,
-    hidden_size=hidden_size,
-    activation=activation)
+    # initialize solver
+    solver = ppsci.solver.Solver(
+        model,
+        constraint,
+        OUTPUT_DIR,
+        optimizer,
+        None,
+        EPOCHS,
+        ITERS_PER_EPOCH,
+        eval_during_train=True,
+        eval_freq=200,
+        equation=equation,
+        geom=geom,
+        validator=validator,
+        visualizer=visualizer,
+    )
+    # train model
+    solver.train()
+    # evaluate after finished training
+    solver.eval()
+    # visualize prediction after finished training
+    solver.visualize()
 
-# Loss
-loss = psci.loss.L2()
-
-# Algorithm
-algo = psci.algorithm.PINNs(net=net, loss=loss)
-
-# Optimizer
-opt = psci.optimizer.Adam(
-    learning_rate=learning_rate, parameters=net.parameters())
-
-# Solver
-solver = psci.solver.Solver(pde=pde_disc, algo=algo, opt=opt)
-solution = solver.solve(num_epoch=epochs)
-
-psci.visu.save_vtk(
-    filename=vtk_filename, geo_disc=pde_disc.geometry, data=solution)
-
-psci.visu.save_npy(
-    filename=solution_filename, geo_disc=pde_disc.geometry, data=solution)
-
-# MSE
-# TODO: solution array to dict: interior, bc
-cord = pde_disc.geometry.interior
-ref = ref_sol(cord[:, 0], cord[:, 1])
-mse2 = np.linalg.norm(solution[0][:, 0] - ref, ord=2)**2
-
-n = 1
-for cord in pde_disc.geometry.boundary.values():
-    ref = ref_sol(cord[:, 0], cord[:, 1])
-    mse2 += np.linalg.norm(solution[n][:, 0] - ref, ord=2)**2
-    n += 1
-
-mse = mse2 / npoints
-
-print("MSE is: ", mse)
+    # directly evaluate pretrained model(optional)
+    logger.init_logger("ppsci", f"{OUTPUT_DIR}/eval.log", "info")
+    solver = ppsci.solver.Solver(
+        model,
+        constraint,
+        OUTPUT_DIR,
+        equation=equation,
+        geom=geom,
+        validator=validator,
+        visualizer=visualizer,
+        pretrained_model_path=f"{OUTPUT_DIR}/checkpoints/latest",
+    )
+    solver.eval()
+    # visualize prediction for pretrained model(optional)
+    solver.visualize()
