--- examples/kovasznay/README.md
@@ -1,31 +1,0 @@
-# Kovasznay Flow #
-
-This is a demo of Kovasznay flow contributing by "XYM-1" in PR-17.
-
-The Navier-Stokes equation describing Kovasznay flows is:
-
-<div align="center">
-<img src="../../docs/source/img/koveq.png" width = "300" align=center />
-</div>
-
-The analytical solution is 
-
-<div align="center">
-<img src="../../docs/source/img/kovsol.png" width = "300" align=center />
-</div>
-
-where the parameters are
-
-<div align="center">
-<img src="../../docs/source/img/kovparam.png" width = "240" align=center />
-</div>
-
-The computation domain is [-0.5,1.5] x [-0.5,1.5]. The simulation results of u and v are as following
-
-<div align="center">
-<img src="../../docs/source/img/kovrstu.png" width = "400" align=center />
-</div>
-
-<div align="center">
-<img src="../../docs/source/img/kovrstv.png" width = "400" align=center />
-</div>
