+++ examples/laplace/laplace2d.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
+# Copyright (c) 2023 PaddlePaddle Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,113 +12,144 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import paddlescience as psci
+# This code is based on PaddleScience/ppsci API
 import numpy as np
-import paddle
 
-cfg = psci.utils.parse_args()
+import ppsci
+from ppsci.utils import config
+from ppsci.utils import logger
 
-if cfg is not None:
-    # Geometry
-    npoints = cfg['Geometry']['npoints']
-    seed_num = cfg['Geometry']['seed']
-    sampler_method = cfg['Geometry']['sampler_method']
-    # Network
-    epochs = cfg['Global']['epochs']
-    num_layers = cfg['Model']['num_layers']
-    hidden_size = cfg['Model']['hidden_size']
-    activation = cfg['Model']['activation']
-    # Optimizer
-    learning_rate = cfg['Optimizer']['lr']['learning_rate']
-    # Post-processing
-    solution_filename = cfg['Post-processing']['solution_filename']
-    vtk_filename = cfg['Post-processing']['vtk_filename']
-    checkpoint_path = cfg['Post-processing']['checkpoint_path']
-else:
-    # Geometry
-    npoints = 10201
-    seed_num = 1
-    sampler_method = 'uniform'
-    # Network
-    epochs = 20000
-    num_layers = 5
-    hidden_size = 20
-    activation = 'tanh'
-    # Optimizer
-    learning_rate = 0.001
-    # Post-processing
-    solution_filename = 'output_laplace2d'
-    vtk_filename = 'output_laplace2d'
-    checkpoint_path = 'checkpoints'
+if __name__ == "__main__":
+    args = config.parse_args()
+    # set random seed for reproducibility
+    ppsci.utils.misc.set_random_seed(42)
+    # set training hyper-parameters
+    EPOCHS = 20000 if not args.epochs else args.epochs
+    ITERS_PER_EPOCH = 1
+    EVAL_FREQ = 200
 
-paddle.seed(seed_num)
-np.random.seed(seed_num)
+    # set output directory
+    OUTPUT_DIR = "./output/laplace2d" if not args.output_dir else args.output_dir
+    logger.init_logger("ppsci", f"{OUTPUT_DIR}/train.log", "info")
 
-# analytical solution 
-ref_sol = lambda x, y: np.cos(x) * np.cosh(y)
+    # set model
+    model = ppsci.arch.MLP(("x", "y"), ("u",), 5, 20)
 
-# set geometry and boundary
-geo = psci.geometry.Rectangular(origin=(0.0, 0.0), extent=(1.0, 1.0))
-geo.add_boundary(
-    name="around",
-    criteria=lambda x, y: (y == 1.0) | (y == 0.0) | (x == 0.0) | (x == 1.0))
+    # set equation
+    equation = {"laplace": ppsci.equation.pde.Laplace(dim=2)}
 
-# discretize geometry
-geo_disc = geo.discretize(npoints=npoints, method=sampler_method)
+    # set geometry
+    geom = {"rect": ppsci.geometry.Rectangle((0.0, 0.0), (1.0, 1.0))}
 
-# Laplace
-pde = psci.pde.Laplace(dim=2)
+    # compute ground truth function
+    def u_solution_func(out):
+        """compute ground truth for u as label data"""
+        x, y = out["x"], out["y"]
+        return np.cos(x) * np.cosh(y)
 
-# set bounday condition
-bc_around = psci.bc.Dirichlet('u', rhs=ref_sol)
+    # set train dataloader config
+    train_dataloader_cfg = {
+        "dataset": "IterableNamedArrayDataset",
+        "iters_per_epoch": ITERS_PER_EPOCH,
+    }
 
-# add bounday and boundary condition
-pde.add_bc("around", bc_around)
+    NPOINT_INTERIOR = 99**2
+    NPOINT_BC = 400
+    NPOINT_TOTAL = NPOINT_INTERIOR + NPOINT_BC
 
-# discretization pde
-pde_disc = pde.discretize(geo_disc=geo_disc)
+    # set constraint
+    pde_constraint = ppsci.constraint.InteriorConstraint(
+        equation["laplace"].equations,
+        {"laplace": 0},
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_TOTAL},
+        ppsci.loss.MSELoss("sum"),
+        evenly=True,
+        name="EQ",
+    )
+    bc = ppsci.constraint.BoundaryConstraint(
+        {"u": lambda out: out["u"]},
+        {"u": u_solution_func},
+        geom["rect"],
+        {**train_dataloader_cfg, "batch_size": NPOINT_BC},
+        ppsci.loss.MSELoss("sum"),
+        criteria=lambda x, y: np.isclose(x, 0.0)
+        | np.isclose(x, 1.0)
+        | np.isclose(y, 0.0)
+        | np.isclose(y, 1.0),
+        name="BC",
+    )
+    # wrap constraints together
+    constraint = {
+        pde_constraint.name: pde_constraint,
+        bc.name: bc,
+    }
 
-# Network
-# TODO: remove num_ins and num_outs
-net = psci.network.FCNet(
-    num_ins=2,
-    num_outs=1,
-    num_layers=num_layers,
-    hidden_size=hidden_size,
-    activation=activation)
+    # set optimizer
+    optimizer = ppsci.optimizer.Adam(learning_rate=0.001)((model,))
 
-# Loss
-loss = psci.loss.L2()
+    # set validator
+    mse_metric = ppsci.validate.GeometryValidator(
+        {"u": lambda out: out["u"]},
+        {"u": u_solution_func},
+        geom["rect"],
+        {
+            "dataset": "IterableNamedArrayDataset",
+            "total_size": NPOINT_TOTAL,
+        },
+        ppsci.loss.MSELoss(),
+        evenly=True,
+        metric={"MSE": ppsci.metric.MSE()},
+        with_initial=True,
+        name="MSE_Metric",
+    )
+    validator = {mse_metric.name: mse_metric}
 
-# Algorithm
-algo = psci.algorithm.PINNs(net=net, loss=loss)
+    # set visualizer(optional)
+    vis_points = geom["rect"].sample_interior(NPOINT_TOTAL, evenly=True)
+    visualizer = {
+        "visulzie_u": ppsci.visualize.VisualizerVtu(
+            vis_points,
+            {"u": lambda d: d["u"]},
+            num_timestamps=1,
+            prefix="result_u",
+        )
+    }
 
-# Optimizer
-opt = psci.optimizer.Adam(
-    learning_rate=learning_rate, parameters=net.parameters())
+    # initialize solver
+    solver = ppsci.solver.Solver(
+        model,
+        constraint,
+        OUTPUT_DIR,
+        optimizer,
+        epochs=EPOCHS,
+        iters_per_epoch=ITERS_PER_EPOCH,
+        eval_during_train=True,
+        eval_freq=EVAL_FREQ,
+        equation=equation,
+        geom=geom,
+        validator=validator,
+        visualizer=visualizer,
+    )
+    # train model
+    solver.train()
+    # evaluate after finished training
+    solver.eval()
+    # visualize prediction after finished training
+    solver.visualize()
 
-# Solver
-solver = psci.solver.Solver(pde=pde_disc, algo=algo, opt=opt)
-solution = solver.solve(num_epoch=epochs)
-
-psci.visu.save_vtk(
-    filename=vtk_filename, geo_disc=pde_disc.geometry, data=solution)
-
-psci.visu.save_npy(
-    filename=solution_filename, geo_disc=pde_disc.geometry, data=solution)
-
-# MSE
-# TODO: solution array to dict: interior, bc
-cord = pde_disc.geometry.interior
-ref = ref_sol(cord[:, 0], cord[:, 1])
-mse2 = np.linalg.norm(solution[0][:, 0] - ref, ord=2)**2
-
-n = 1
-for cord in pde_disc.geometry.boundary.values():
-    ref = ref_sol(cord[:, 0], cord[:, 1])
-    mse2 += np.linalg.norm(solution[n][:, 0] - ref, ord=2)**2
-    n += 1
-
-mse = mse2 / npoints
-
-print("MSE is: ", mse)
+    # directly evaluate model from pretrained_model_path(optional)
+    logger.init_logger("ppsci", f"{OUTPUT_DIR}/eval.log", "info")
+    solver = ppsci.solver.Solver(
+        model,
+        constraint,
+        OUTPUT_DIR,
+        equation=equation,
+        geom=geom,
+        validator=validator,
+        visualizer=visualizer,
+        pretrained_model_path=f"{OUTPUT_DIR}/checkpoints/latest",
+    )
+    solver.eval()
+    # visualize prediction from pretrained_model_path(optional)
+    solver.visualize()
